pyproject.toml
[tool.poetry]
name = "rttptool"
version = "0.1.0"
description = "Role Templates Testing Protocol (RTTP) reference inplementation."
authors = ["Alexey Busygin <yaabusygin@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
Jinja2 = "^2.11"
packaging = "^20.9"

[tool.poetry.dev-dependencies]
importlib_resources = { version = "^5.1", python = "<3.7" }
PyYAML = "^5.4"
pylint = "^2.8"
pydocstyle = "^6.0"
pycodestyle = "^2.7"
mypy = "0.812"

[tool.pylint."MESSAGES CONTROL"]
disable = [
    # prefer pycodestyle
    "line-too-long",
    "unnecessary-semicolon",
    "trailing-whitespace",
    "missing-final-newline",
    "bad-indentation",
    "multiple-statements",
    "bare-except",

    # prefer pydocstyle
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",

    # prefer mypy
    "syntax-error",

    # actually disable
    "too-few-public-methods",
    "duplicate-code"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
